requirements.txt
flask==3.1.0
gunicorn==21.2.0
APScheduler==3.10.4
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
